Move some dom manipulation to domUpdate file
